fixes #1809: Trying some refatcoring using proxy/kerberos as a mixin
