+++ lib/proxy/dns/nsupdate_gss.rb
@@ -1,8 +1,11 @@
 require 'proxy/dns/nsupdate'
-require 'rkerberos'
+require 'proxy/kerberos'
 
 module Proxy::DNS
   class NsupdateGSS < Nsupdate
+
+    include Proxy::Kerberos
+
     attr_reader :tsig_keytab, :tsig_principal
 
     def initialize options = {}
@@ -25,19 +28,19 @@ module Proxy::DNS
       super
     end
 
-    private
-
-    def init_krb5_ccache
-      krb5 = Kerberos::Krb5.new
-      ccache = Kerberos::Krb5::CredentialsCache.new
-      logger.info "Requesting credentials for Kerberos principal #{tsig_principal} using keytab #{tsig_keytab}"
-      begin
-        krb5.get_init_creds_keytab tsig_principal, tsig_keytab, nil, ccache
-      rescue => e
-        logger.error "Failed to initialise credential cache from keytab: #{e}"
-        raise Proxy::DNS::Error.new("Unable to initialise Kerberos: #{e}")
-      end
-      logger.debug "Kerberos credential cache initialised with principal: #{ccache.primary_principal}"
-    end
+#    private
+#
+#    def init_krb5_ccache
+#      krb5 = Kerberos::Krb5.new
+#      ccache = Kerberos::Krb5::CredentialsCache.new
+#      logger.info "Requesting credentials for Kerberos principal #{tsig_principal} using keytab #{tsig_keytab}"
+#      begin
+#        krb5.get_init_creds_keytab tsig_principal, tsig_keytab, nil, ccache
+#      rescue => e
+#        logger.error "Failed to initialise credential cache from keytab: #{e}"
+#        raise Proxy::DNS::Error.new("Unable to initialise Kerberos: #{e}")
+#      end
+#      logger.debug "Kerberos credential cache initialised with principal: #{ccache.primary_principal}"
+#    end
   end
 end
